Create a test file for the ast_builder (incomplete)
[ci skip]
